fix(accompagnement): privilegier le code commune pour rechercher les accompagnements sur les grandes villes
