+++ src/pages/api/etablissements-accompagnement/index.controller.test.ts
@@ -19,7 +19,7 @@ describe('rechercher un établissement d‘accompagnement', () => {
 	describe('lorsque la recherche est valide', () => {
 		it('retourne la liste des établissements d‘accompagnement', async () => {
 			nock('https://api-lannuaire.service-public.fr/api/explore/v2.1')
-				.get('/catalog/datasets/api-lannuaire-administration/records?where=suggest(adresse,%22code_postal%2034001%22)and%20pivot%20LIKE%20%22cij%22&limit=100&select=adresse,telephone,adresse_courriel,nom,id,pivot,plage_ouverture')
+				.get('/catalog/datasets/api-lannuaire-administration/records?where=suggest(pivot,%22code_insee_commune%2034000%22)and%20pivot%20LIKE%20%22cij%22&limit=100&select=adresse,telephone,adresse_courriel,nom,id,pivot,plage_ouverture')
 				.reply(200, aResultatRechercheEtablissementPublicListResponse());
 
 			await testApiHandler<EtablissementAccompagnement[] | ErrorHttpResponse>({
@@ -29,7 +29,7 @@ describe('rechercher un établissement d‘accompagnement', () => {
 					const json = await res.json();
 					expect(json).toEqual(anEtablissementAccompagnementList());
 				},
-				url: '/etablissements-accompagnement?codePostal=34001&codeCommune=34000&typeAccompagnement=cij',
+				url: '/etablissements-accompagnement?codeCommune=34000&typeAccompagnement=cij',
 			});
 		});
 	});
@@ -46,7 +46,7 @@ describe('rechercher un établissement d‘accompagnement', () => {
 					const json = await res.json();
 					expect(json).toEqual({ error: ErreurMetier.CONTENU_INDISPONIBLE });
 				},
-				url: '/etablissements-accompagnement?codePostal=34001&codeCommune=34000&typeAccompagnement=cij',
+				url: '/etablissements-accompagnement?codeCommune=34000&typeAccompagnement=cij',
 			});
 		});
 	});
