+++ src/pages/api/etablissements-accompagnement/index.controller.ts
@@ -13,7 +13,6 @@ import { dependencies } from '~/server/start';
 
 export const querySchema = Joi.object({
 	codeCommune: Joi.string().alphanum().max(5).required(),
-	codePostal: Joi.string().alphanum().max(5).required(),
 	typeAccompagnement: Joi.string().valid(...Object.values(TypeÉtablissement)).required(),
 }).options({ allowUnknown: true });
 
@@ -24,7 +23,7 @@ export async function rechercherÉtablissementAccompagnementHandler(
 	const résultatsRechercheÉtablissementAccompagnement = await dependencies
 		.établissementAccompagnementDependencies
 		.rechercherEtablissementAccompagnementUseCase
-		.handle({ codeCommune: String(query.codeCommune), codePostal: String(query.codePostal), typeAccompagnement: String(query.typeAccompagnement) });
+		.handle({ codeCommune: String(query.codeCommune), typeAccompagnement: String(query.typeAccompagnement) });
 	return handleResponse(résultatsRechercheÉtablissementAccompagnement, res);
 }
