+++ src/server/etablissement-accompagnement/infra/apiEtablissementPublic.repository.test.ts
@@ -29,11 +29,10 @@ describe('ApiÉtablissementPublicRepository', () => {
 			it('lorsque je fais une recherche avec la ville de Marseille, appelle l‘api avec l‘ensemble des arrondissement de Marseille', async () => {
 				const httpClient = aPublicHttpClientService();
 				const repository = new ApiEtablissementPublicRepository(httpClient, anErrorManagementService());
-				const codePostal = '13013';
 				const codeCommune = '13055';
 				const typeAccompagnement = TypeÉtablissement.INFO_JEUNE;
 
-				await repository.search({ codeCommune, codePostal, typeAccompagnement });
+				await repository.search({ codeCommune, typeAccompagnement });
 
 				expect(httpClient.get).toHaveBeenCalledWith(expect.stringContaining('suggest(adresse,%22code_postal%20130*%22)'));
 			});
@@ -41,11 +40,10 @@ describe('ApiÉtablissementPublicRepository', () => {
 			it('lorsque je fais une recherche avec la ville de Lyon, appelle l‘api avec l‘ensemble des arrondissement de Lyon', async () => {
 				const httpClient = aPublicHttpClientService();
 				const repository = new ApiEtablissementPublicRepository(httpClient, anErrorManagementService());
-				const codePostal = '69002';
 				const codeCommune = '69123';
 				const typeAccompagnement = TypeÉtablissement.INFO_JEUNE;
 
-				await repository.search({ codeCommune, codePostal, typeAccompagnement });
+				await repository.search({ codeCommune, typeAccompagnement });
 
 				expect(httpClient.get).toHaveBeenCalledWith(expect.stringContaining('suggest(adresse,%22code_postal%206900*%22)'));
 			});
@@ -53,26 +51,24 @@ describe('ApiÉtablissementPublicRepository', () => {
 			it('lorsque je fais une recherche avec la ville de Paris, appelle l‘api avec l‘ensemble des arrondissement de Paris', async () => {
 				const httpClient = aPublicHttpClientService();
 				const repository = new ApiEtablissementPublicRepository(httpClient, anErrorManagementService());
-				const codePostal = '75002';
 				const codeCommune = '75056';
 				const typeAccompagnement = TypeÉtablissement.INFO_JEUNE;
 
-				await repository.search({ codeCommune, codePostal, typeAccompagnement });
+				await repository.search({ codeCommune, typeAccompagnement });
 
 				expect(httpClient.get).toHaveBeenCalledWith(expect.stringContaining('suggest(adresse,%22code_postal%20750*%22)'));
 			});
 		});
 
-		it('lorsque l‘utilisateur recherche une ville sans arrondissement, fait la recherche avec le code postal', async () => {
+		it('lorsque l‘utilisateur recherche une ville sans arrondissement, fait la recherche avec le code commune insee', async () => {
 			const httpClient = aPublicHttpClientService();
 			const repository = new ApiEtablissementPublicRepository(httpClient, anErrorManagementService());
-			const codePostal = '83400';
 			const codeCommune = '83000';
 			const typeAccompagnement = TypeÉtablissement.INFO_JEUNE;
 
-			await repository.search({ codeCommune, codePostal, typeAccompagnement });
+			await repository.search({ codeCommune, typeAccompagnement });
 
-			expect(httpClient.get).toHaveBeenCalledWith(expect.stringContaining(`suggest(adresse,%22code_postal%20${codePostal}%22)`));
+			expect(httpClient.get).toHaveBeenCalledWith(expect.stringContaining(`suggest(pivot,%22code_insee_commune%20${codeCommune}%22)`));
 		});
 		describe('lorsque la recherche retourne une 200', () => {
 			it('retourne la liste des établissements d‘accompagnement', async () => {
@@ -83,16 +79,15 @@ describe('ApiÉtablissementPublicRepository', () => {
 					.mockResolvedValue(anAxiosResponse(aResultatRechercheEtablissementPublicListResponse()));
 				const repository = new ApiEtablissementPublicRepository(httpClient, anErrorManagementService());
 				const expected = createSuccess(anEtablissementAccompagnementList());
-				const codePostal = '46100';
 				const codeCommune = '46000';
 				const typeAccompagnement = TypeÉtablissement.INFO_JEUNE;
 
 				// when
-				const result = await repository.search({ codeCommune, codePostal, typeAccompagnement });
+				const result = await repository.search({ codeCommune, typeAccompagnement });
 
 				// then
 				expect(httpClient.get).toHaveBeenCalledWith(expect.stringContaining('catalog/datasets/api-lannuaire-administration/records?'));
-				expect(httpClient.get).toHaveBeenCalledWith(expect.stringContaining('where=suggest(adresse,%22code_postal%2046100%22)and%20pivot%20LIKE%20%22cij%22'));
+				expect(httpClient.get).toHaveBeenCalledWith(expect.stringContaining('where=suggest(pivot,%22code_insee_commune%2046000%22)and%20pivot%20LIKE%20%22cij%22'));
 				expect(httpClient.get).toHaveBeenCalledWith(expect.stringContaining('&limit=100&select=adresse,telephone,adresse_courriel,nom,id,pivot,plage_ouverture'));
 				expect(result).toEqual(expected);
 			});
@@ -108,13 +103,12 @@ describe('ApiÉtablissementPublicRepository', () => {
 				const errorManagementService = anErrorManagementService({
 					handleFailureError: jest.fn(() => createFailure(expectedError)),
 				});
-				const codePostal = '46100';
 				const codeCommune = '46000';
 				const typeAccompagnement = 'cij';
 				const repository = new ApiEtablissementPublicRepository(httpClient, errorManagementService);
 
 				// when
-				const result = await repository.search({ codeCommune, codePostal, typeAccompagnement });
+				const result = await repository.search({ codeCommune, typeAccompagnement });
 
 				// then
 				expect(errorManagementService.handleFailureError).toHaveBeenCalledWith(httpError, logInformation);
@@ -136,7 +130,7 @@ describe('ApiÉtablissementPublicRepository', () => {
 				const repository = new ApiEtablissementPublicRepository(httpClientService, errorManagementServiceSearch);
 
 				// When
-				const result = await repository.search({ codeCommune: '46000', codePostal: '46100', typeAccompagnement: 'cij' });
+				const result = await repository.search({ codeCommune: '46000', typeAccompagnement: 'cij' });
 
 				// Then
 				expect(result.instance).toEqual('success');
@@ -176,7 +170,7 @@ describe('ApiÉtablissementPublicRepository', () => {
 				const repository = new ApiEtablissementPublicRepository(httpClientService, errorManagementServiceSearch);
 
 				// When
-				const result = await repository.search({ codeCommune: '46000', codePostal: '46100', typeAccompagnement: 'cij' });
+				const result = await repository.search({ codeCommune: '46000', typeAccompagnement: 'cij' });
 
 				// Then
 				expect(errorManagementServiceSearch.logValidationError).not.toHaveBeenCalled();
